+++ CHANGELOG.md
@@ -8,5 +8,6 @@
    * Added support for StorageClass - https://github.com/fabric8io/kubernetes-client/pull/978
 
   Improvements
-
+   * Fixed issue of SecurityContextConstraints not working - https://github.com/fabric8io/kubernetes-client/pull/982
+	Note :- This got fixed by fixing model - https://github.com/fabric8io/kubernetes-model/pull/274
   Dependencies Upgrade